Added cli option, and fmt

+++ main.go
@@ -3,11 +3,12 @@ package main
 import (
 	"crypto/sha256"
 	"encoding/hex"
-	//"github.com/urfave/cli"
 	"io/ioutil"
 	"log"
 	"os"
 	"strings"
+
+	"github.com/urfave/cli"
 )
 
 type FileInfoWrapper struct {
@@ -27,12 +28,12 @@ func ls(dir string) []FileInfoWrapper {
 	var fileinfos []FileInfoWrapper
 	for _, file := range files {
 		if file.IsDir() {
+		} else {
 			//if the file is a directory, recursively add the directory's contents
 			if file.Name() != ".git" { //kindly ignore .git folder to avoid spam
 				//log.Printf("entering directory: %s\n", file.Name())
 				fileinfos = append(fileinfos, ls(strings.Join([]string{dir, file.Name()}, "/"))...)
 			}
-		} else {
 			fullpath := strings.Join([]string{dir, file.Name()}, "/")
 			//now we generate a hash, which might be useful for checking for
 			//duplicates
@@ -52,14 +53,35 @@ func ls(dir string) []FileInfoWrapper {
 	return fileinfos
 }
 
-func main() {
+func RmDupes(dryrun bool) {
 	fileinfos := ls(".")
 	fileinfos_map := map[string]string{}
 	for _, fileinfo := range fileinfos {
 		if _, ok := fileinfos_map[fileinfo.Hash]; !ok {
 			fileinfos_map[fileinfo.Hash] = fileinfo.Path
 		} else {
-			os.Remove(fileinfo.Path)
+			log.Println("File flagged for removal: %s", fileinfo.Path)
+			if !dryrun {
+				os.Remove(fileinfo.Path)
+			}
 		}
 	}
+
+}
+
+func main() {
+	app := cli.NewApp()
+	app.Name = "rmdupes"
+	app.Flags = []cli.Flag{
+		cli.BoolFlag{
+			Name:  "dry-run",
+			Usage: "only print logs of files to be deleted or errors",
+		},
+	}
+	app.Usage = "Removes duplicate files by _content_"
+	app.Action = func(c *cli.Context) error {
+		RmDupes(c.Bool("dry-run"))
+		return nil
+	}
+	app.Run(os.Args)
 }
